+++ docs/docusaurus.config.js
@@ -1,8 +1,8 @@
 // @ts-check
 // Note: type annotations allow type checking and IDEs autocompletion
 
-const lightCodeTheme = require('./src/theme/LightCustomTheme.js');
-const darkCodeTheme = require('./src/theme/DarkCustomTheme.js');
+const lightCodeTheme = require('./src/theme/lightCustomTheme.js')
+const darkCodeTheme = require('./src/theme/darkCustomTheme.js')
 /** @type {import('@docusaurus/types').Config} */
 const config = {
   title: 'React Native Ficus UI',
